Move submission period logic to its own module
