+++ feedback/api.py
@@ -2,4 +2,4 @@
 from feedback.league import create_league
 from feedback.league import get_league
 
-from feedback.submit import create_submission
+from feedback.submission import create_submission
